Phase 2: H618 platform abstraction layer (T006-T014)
- T006: H618 platform detection via /proc/device-tree/compatible
- T007: ORANGEPI_H618_PIO_BASE = 0x0300B000 constant
- T008: lib/gpio-bits-h618.h with H618 GPIO register offsets
- T009: Platform-aware mmap and GPIO::Init for H618 base address
- T010: H618 GPIO direction configuration via port config registers
- T011: H618 GPIO set/clear/read with virtual pin translation
- T012: orangepi-zero2w mapping struct in hardware-mapping.c
- T013: Registered orangepi-zero2w in kHardwareMappings array
- T014: Updated --led-gpio-mapping help text

Key design: Uses virtual pin indices 0-13 in HardwareMapping to avoid
64-bit GPIO number overflow. kH618PinTable translates virtual indices
to actual H618 (port, pin) pairs at runtime.

gpio_bits_t changed to uint64_t unconditionally for H618 support.
Clean build with zero warnings.

+++ lib/gpio-bits.h
@@ -19,10 +19,15 @@
 #define RPI_GPIOBITS_H
 
 #include <stdint.h>
+
+// ORANGE_PI_ZERO2W: Always use 64-bit to support H618 GPIO numbers (0-287).
+// The original code used uint32_t for BCM (GPIO 0-31) and uint64_t for
+// compute-module. H618 uses GPIO numbers 226-271, requiring at least
+// 272 bits... but we use a virtual mapping (indices 0-13) so 64-bit suffices.
 #ifdef ENABLE_WIDE_GPIO_COMPUTE_MODULE
 typedef uint64_t gpio_bits_t;
 #else
-typedef uint32_t gpio_bits_t;
+typedef uint64_t gpio_bits_t;
 #endif
 
 #endif
